Python/PyQt5/requirements.txt
cssutils
cycler
fonttools
GitPython
importlib_resources
kiwisolver
matplotlib
numpy
packaging
Pillow
pyparsing
PyQt6==6.2.3
PyQt6-Qt6==6.2.4
PyQt6-sip==13.2.1
PyQt6-WebEngine==6.2.1
PyQt6-WebEngine-Qt6==6.2.4
python-dateutil
six
